# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LabelHub API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "创建任务",
        "description": "createTask 命令（契约 §18.1）：创建 DRAFT 状态任务，写入 TASK_CREATED audit log。",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "获取任务列表（OWNER）",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          },
          {
            "name": "includeStats",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includestats"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "查询任务详情",
        "description": "获取任务详情（契约 §23.1：OWNER/REVIEWER/LABELER 均可访问）。",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "更新任务（仅 DRAFT 状态）",
        "description": "更新 DRAFT 任务字段（契约 §23.1）。非 DRAFT 状态返回 409。",
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "删除草稿任务（仅 DRAFT）",
        "description": "删除草稿任务，级联清理 schema 草稿 / 版本 / 数据项 / 审计日志。非 DRAFT 返回 409。",
        "operationId": "delete_draft_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/stats": {
      "get": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "任务概览统计（OWNER 看板：进度 / 各状态计数 / 剩余配额）",
        "operationId": "get_task_stats_api_v1_tasks__task_id__stats_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/publish": {
      "post": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "发布任务",
        "description": "publishTask 命令（契约 §16 / §18.1）：DRAFT → PUBLISHED，校验 schemaVersionId。",
        "operationId": "publish_task_api_v1_tasks__task_id__publish_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/pause": {
      "post": {
        "tags": [
          "tasks",
          "schema"
        ],
        "summary": "暂停任务",
        "description": "pauseTask 命令（契约 §18.1）：PUBLISHED → PAUSED。",
        "operationId": "pause_task_api_v1_tasks__task_id__pause_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskResponse": {
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskResponse"
          },
          "auditLog": {
            "$ref": "#/components/schemas/AuditLogSummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "task",
          "auditLog"
        ],
        "title": "CreateTaskResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseTaskRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "PauseTaskRequest"
      },
...
      "PublishTaskRequest": {
        "properties": {
          "schemaVersionId": {
            "type": "string",
            "title": "Schemaversionid",
            "description": "待激活的 SchemaVersion ID"
          },
          "reviewConfigId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewconfigid",
            "description": "AI 审核配置 ID，可选"
          },
          "reviewDisabledExplicitly": {
            "type": "boolean",
            "title": "Reviewdisabledexplicitly",
            "description": "显式声明不使用 AI 审核",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "schemaVersionId"
        ],
        "title": "PublishTaskRequest",
        "description": "publishTask 命令入参（契约 §16 / §18.1）"
      },
      "PublishTaskResponse": {
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskResponse"
          },
          "auditLog": {
            "$ref": "#/components/schemas/AuditLogSummaryResponse"
          }
        },
        "type": "object",
        "required": [
          "task",
          "auditLog"
        ],
        "title": "PublishTaskResponse"
      },
...
    }
  }
}